+++ src/translator.cc
@@ -352,4 +352,23 @@ namespace ctranslate2 {
     return _model->device();
   }
 
+  int Translator::device_index() const {
+    return _model->device_index();
+  }
+
+  ComputeType Translator::compute_type() const {
+    return _model->compute_type();
+  }
+
+  void Translator::set_model(const std::string& model_dir)
+  {
+    std::shared_ptr<const models::Model> model;
+    set_model(model->load(model_dir, this->device(), this->device_index(), this->compute_type()));
+  }
+
+  void Translator::set_model(const std::shared_ptr<const models::Model>& model)
+  {
+    _model = model;
+    make_graph();
+  }
 }